Updated KISS File
Welp, I ran into a... "couple" issues. In the meantime, here's a simplified version of the file. I should note that this does NOT currently work the way I intend it to, hence the simplification. Once I find the time, I will be properly researching how to work with arrays on here, lol.